Add CF invalidation; update open-next; add readme

+++ packages/open-next-test/.yalc/open-next-conico-fork/build.js
@@ -113,10 +113,22 @@ function createServerBundle(monorepoRoot) {
     //Copy prerendered pages to cache folder
     const buildId = fs.readFileSync(path.join(appPath, ".next/BUILD_ID"), "utf-8");
     const outputCache = path.join(outputDir, "cache", buildId);
-    fs.cpSync(path.join(appPath, ".next/standalone", packagePath, ".next/server/pages"), outputCache, {
-        recursive: true,
-        verbatimSymlinks: true,
-    });
+    //Copy pages to cache folder
+    const pagesPrerenderedPath = path.join(appPath, ".next/standalone", packagePath, ".next/server/pages");
+    if (fs.existsSync(pagesPrerenderedPath)) {
+        fs.cpSync(path.join(appPath, ".next/standalone", packagePath, ".next/server/pages"), outputCache, {
+            recursive: true,
+            verbatimSymlinks: true,
+        });
+    }
+    //Copy app to cache folder
+    const appPrerenderedPath = path.join(appPath, ".next/standalone", packagePath, ".next/server/app");
+    if (fs.existsSync(appPrerenderedPath)) {
+        fs.cpSync(appPrerenderedPath, outputCache, {
+            recursive: true,
+            verbatimSymlinks: true,
+        });
+    }
     const removeJsFiles = (dir) => {
         fs.readdirSync(dir).forEach((file) => {
             if (fs.statSync(path.join(dir, file)).isDirectory()) {
